+++ src/common.rs
@@ -3,9 +3,7 @@ pub(crate) use std::{
   borrow::Cow,
   cmp,
   collections::{BTreeMap, BTreeSet},
-  convert::AsRef,
   env,
-  ffi::OsStr,
   fmt::{self, Display, Formatter},
   fs,
   io::{self, Write},
